+++ metadata.rb
@@ -20,8 +20,9 @@ supports 'ubuntu', '= 12.04'
 supports 'ubuntu', '= 14.04'
 supports 'arch'
 
+recommends 'chef-provisioning'
+
 depends 'libarchive'
-depends 'chef-provisioning'
 depends 'golang', '~> 1.4'
 depends 'runit'
 depends 'yum-repoforge'